converted big ints to longs

+++ src/TaskSet.java
@@ -2,7 +2,9 @@ import java.math.BigInteger;
 
 public class TaskSet {
 	private String owner; // who is responsible for this TaskGroup
-	private BigInteger begin;	  // the first TaskGroup this TaskGroup is responsible for
-	private BigInteger end;	  // the last TaskGroup this TaskGroup is responsible for
+	private long begin;	  // the first TaskGroup this TaskGroup is responsible for
+	private long end;	  // the last TaskGroup this TaskGroup is responsible for
+	
+	
 
 }
